refactor(go-template): extract template-literal memo lowering into memo/template-interp.ts
Move the template-literal memo initial-value lowering out of the adapter
class into pure free functions (Roadmap B):
computeTemplateLiteralMemoInitialValue, resolveTemplateInterpolation,
parseLocalKeyBinding, recordIndexInterpolationToGo, propsAccessName. They
read only state.localConstants / state.propsObjectName / state.usesFmt.

GoEmitContext gains resolveModuleStringConst — the one adapter-resident
entry point this module calls back into (it depends on per-compile loop
state that stays on the adapter). Removing the cluster also reunites
computeMemoInitialValue with its doc comment, which the template method
had separated.

Output byte-identical, verified by the adapter unit (552) and
adapter-tests conformance (786) suites; tsgo + biome clean.
